Moving tests to better match structure
